app/requirements.txt
Flask==2.0.2
#Flask-MySQL
#flask-mysqldb
requests
mysql-connector-python
